test: document configurable property scenarios
